Changed message size lengths from short to long.

+++ thesis_core/src/main/java/thesis/network/messages/SetSimStepRateMsg.java
@@ -38,9 +38,9 @@ public class SetSimStepRateMsg extends InfrastructureMsg
    }
 
    @Override
-   public short getEncodedSize()
+   public long getEncodedSize()
    {
-     return (short)Integer.BYTES;
+     return Integer.BYTES;
    }
 
 }
